feat: 声からnoteの下書きまで繋ぎ、画面に wake 行と操作ボタンを足す
構成の作り直し（挙動は不変）:
- server/app.py 865行→155行。session / hub / approval / facts / scheme /
  gitstate / clock へ分離。app.py には配線とルートだけ残す
- App.tsx 696行→327行。hook 4本と theme / copy / model / clap-settings へ分離
- 停止していた先読みブリーフィング（briefing.py + テスト294行）を削除

機能:
- 声で「note書いて」から下書きまで（router に NOTE / note_writer / thumbnail）
- 1段目に wake 行、3段目に操作ボタン4つ、チャットの打ち込み欄
- 起動を指パッチン1回へ戻す（本人の指定。真値は audio/types.ts）
- 「壁紙」の聞き間違い（壁が見／風が見）を実機ログから合図へ追加

不具合:
- 「続けて聞く」のあと LISTENING から待機へ戻らない（speaking を降ろす場所）
- 見出し画像が入っているのに失敗と報告する／確定ダイアログが保存ボタンを覆う
- 編集画面からプレビューへ流されると下書き保存できない
- 切断時に走っている Codex が止まらない

文書:
- DESIGN.md に §15.2（画面）§18.2（声→note）を追加。§5 の構成表を実物へ
- ビジュアル仕様を §15 に一本化し、旧版を docs/archive/ へ退避
- 消えていた §16「常設運用」の見出しを復元
- docs/2026-09-05-refactor-handover.md（実機で確かめた／確かめていないの別つき）

pytest 125 / vitest 21 / tsc / build 通過。実機は一部のみ確認。

+++ apps/iphone-ui/src/audio/types.ts
@@ -1,4 +1,9 @@
 export interface AudioMetrics { at: number; rms: number; hfRatio: number; riseMs: number }
 export interface ClapLog extends AudioMetrics { accepted: boolean; reason: string }
 export interface ClapSettings { ratio: number; hfMin: number; gapMin: number; gapMax: number; mode: "double" | "single" }
-export const DEFAULT_CLAP_SETTINGS: ClapSettings = { ratio: 6, hfMin: 0.2, gapMin: 250, gapMax: 800, mode: "double" };
+// 起動方式の真値はここ（2026-09-05、本人の指定で "single" へ戻した）。
+// 2026-09-02 に "double" にしたのはタイピング音の誤起動が理由で、
+// 9/5 のログでも single 相当の判定は 14回 → 52回に増える見込みだった。
+// **誤起動が続くようなら ratio / hfMin を上げて対処する。**
+// mode を勝手に戻さないこと（本人の指定が優先）
+export const DEFAULT_CLAP_SETTINGS: ClapSettings = { ratio: 6, hfMin: 0.2, gapMin: 250, gapMax: 800, mode: "single" };
